Added delete button for individual list items
Functionality added to app which allows users to delete individual list items, rather than having to wipe the whole list. The front-end code employs the existing functionality to refresh the list after a deletion operation has been completed, though latency in the REST service means that occasionally, a delete command has not yet been processed by the time the list is refreshed. This will be fixed in the near future by the use of async/await functionality.