Simplified deletion of individual todos
Previously, deleting an individual 'todo' would delete all items from the front end, send a specific request to the back end deleting the denoted todo, and re-run the getAll function to build the front end todo list based on information gathered from the database. Instead, the application now removes the specific item from the front end and from the back end. This removes the architectural issue of over-engineering the function, and also massively reduces the chances of a GET request being completed before the DELETE request (which happened frequently before).