Bug Fix - deletion by index not name
Previously, this application was configured to delete records based on the record's text. This did not deal well with instances whereby there were two or more records containing the same text. Whilst this was not catastrophic, it was lazy and bad practice. The server is now configured to delete records based on a unique id for each item.