Return to default db indexing behaviour
In the last commit, I implemented functionality on the front and back ends of the application whereby 'todo' items were marked with id's reflective of their database index. The database was set up to auto increment, so that each new todo would have a unique index. However, the long-term problem with this solution was that even after deleting db records, the indexing would continue to increase inexorably, placing greater and greater resource load on the system. I have now returned the database to its default behaviour, and as a result, also altered the front-end code so that deletion of individual items is enabled without id assignment.