App configured to retrieve todos on load
I have removed the 'retrieve todos' button, as this was causing issues with phantom todo items. List manipulation in the UI was outpacing communications sent to the backend (because it is slower for the server to react to http requests than for the frontend to manipulate dom elements), hence on numerous occasions, exessive button pressing caused problems! The application now uses an IIFE to fire get requests when the page is loaded.