Alterations to DB Structure
I have altered the database structure as the first part of a re-engineering of deletion and edit functionality. As stated in this stack overflow answer (https://stackoverflow.com/a/35876869), the values saved within rowid can be altered by the vacuum command used by the database. Up until this point, I have been experiencing issues with deletion and edit functionality which I have traced to the database commands on my server-side code (all of which relate to specific rows define by 'rowid'). As such, I have added a specific ID column, which will take numeric inputs to be used as IDs, and serve as a reference identifier for the deletion/edit of specific records.